+++ server/apps/api.go
@@ -6,5 +6,5 @@ package apps
 type API interface {
 	Call(Call) (*CallResponse, error)
 	InstallApp(InInstallApp) (*OutInstallApp, error)
-	GetWidgets(userID, channelID string) ([]LocationInt, error)
+	GetLocations(userID, channelID string) ([]LocationInt, error)
 }
